--- modules/react-components/src/components/ai/index.ts
@@ -1,3 +1,0 @@
-
-export * from "./ai-card";
-export * from "./loading-screen";
