+++ modules/react-components/src/components/index.ts
@@ -64,4 +64,3 @@ export * from "./typography";
 export * from "./transfer";
 export * from "./redux-forms";
 export * from "./validation";
-export * from "./ai";
